Add time series tests

+++ src/util.jl
@@ -75,11 +75,11 @@ function elevationangle(times::Vector{Float64}, phasedir::Array{Float64,2}, stat
 end
 
 """
-    gentimeseries!(series::Vector{ComplexF32}, mode::String, location::ComplexF32, scale::Float64, driftrate::Float64, nsamples::Int64, rng::AbstractRNG)
+    gentimeseries!(series::Vector{ComplexF32}, mode::String, location::ComplexF32, scale::Float32, driftrate::Float32, nsamples::Int64, rng::AbstractRNG)
 
 Generate a complex-valued Gaussian process time-series of length nsamples with the given location, scale, and driftrate parameters.
 """
-function gentimeseries!(series::Vector{ComplexF32}, mode::String, location::ComplexF32, scale::Float64, driftrate::Float64, nsamples::Int64, rng::AbstractRNG)
+function gentimeseries!(series::Vector{ComplexF32}, mode::String, location::ComplexF32, scale::Float32, driftrate::Float32, nsamples::Int64, rng::AbstractRNG)
     # TODO this is a crude version of a wiener process -- to be updated
     if mode == "gp"
         sqrtnsamples = sqrt(nsamples)
@@ -94,11 +94,11 @@ function gentimeseries!(series::Vector{ComplexF32}, mode::String, location::Comp
 end
 
 """
-    gentimeseries!(series::Vector{Float32}, mode::String, location::Float32, scale::Float64, driftrate::Float64, nsamples::Int64, rng::AbstractRNG)
+    gentimeseries!(series::Vector{Float32}, mode::String, location::Float32, scale::Float32, driftrate::Float32, nsamples::Int64, rng::AbstractRNG)
 
 Generate a complex-valued Gaussian process time-series of length nsamples with the given location, scale, and driftrate parameters.
 """
-function gentimeseries!(series::Vector{Float32}, mode::String, location::Float32, scale::Float64, driftrate::Float64, nsamples::Int64, rng::AbstractRNG)
+function gentimeseries!(series::Vector{Float32}, mode::String, location::Float32, scale::Float32, driftrate::Float32, nsamples::Int64, rng::AbstractRNG)
     # TODO this is a crude version of a wiener process -- to be updated
     if mode == "gp"
         sqrtnsamples = sqrt(nsamples)
